Move some utils that use get_proper_type away from types.py
